[global] added additional parameters to configure access to registry (CA, scheme)

See merge request sys/deckhouse-oss!517

+++ candi/bashible/bundles/centos-7/all/000_install_ca_certificates.sh.tpl
@@ -14,4 +14,17 @@
 
 bb-yum-install ca-certificates
 # hack to avoid problems with certs in alpine busybox for kube-apiserver
-ln -s /etc/ssl/certs/ca-bundle.crt /etc/ssl/certs/ca-certificates.crt
+if [[ ! -e /etc/ssl/certs/ca-certificates.crt ]]; then
+  ln -s /etc/ssl/certs/ca-bundle.crt /etc/ssl/certs/ca-certificates.crt
+fi
+
+{{- if .registry.ca }}
+bb-event-on 'registry-ca-changed' '_update_ca_certificates'
+function _update_ca_certificates() {
+  update-ca-trust
+}
+
+bb-sync-file /etc/pki/ca-trust/source/anchors/registry-ca.crt - registry-ca-changed << "EOF"
+{{ .registry.ca }}
+EOF
+{{- end }}
